+++ lib/postgrex/agtype_extension.ex
@@ -12,28 +12,27 @@ defmodule AshAge.Postgrex.AgtypeExtension do
   def format(_), do: :text
 
   def encode(_) do
-    quote do
+    quote location: :keep do
       value when is_binary(value) ->
-        value
+        [<<byte_size(value)::int32()>> | value]
 
-      _ ->
-        :error
+      other ->
+        raise DBConnection.EncodeError,
+              Postgrex.Utils.encode_msg(other, "a binary")
     end
   end
 
   def decode(:copy) do
-    quote do
-      value when is_binary(value) ->
+    quote location: :keep do
+      <<len::int32(), value::binary-size(len)>> ->
         :binary.copy(value)
-
-      value ->
-        value
     end
   end
 
   def decode(_) do
-    quote do
-      value -> value
+    quote location: :keep do
+      <<len::int32(), value::binary-size(len)>> ->
+        value
     end
   end
 end